APIs/Payment/requirements.txt
fastapi
uvicorn
stripe
mysql-connector-python
pydantic

python-dotenv
